refactor(scan): include `SCAN_VERSION` when saving
include the current `SCAN_VERSION` in the saved scan record file. this
way you actually have something to compare against when it's eventually
updated.

just stores the u16 as the first two bytes of the file. since the file is
no longer fully plaintext, removes the `.json` extension as well - which
will be even more relevant for when i get around to replacing it with
rkyv or bitcode.

also includes a few cleanups for scan thread code.