feat: adjust tracing directives and log formatting
makes it so useless symphonia/blade/zbus spam is hidden by default

the way env vars are checked isn't beautiful but because of the design of
`EnvFilter` and its builder api, this is really the only option to get
what i want. with_default_directive doesnt let you specify multiple
directives (like the whole point of using EnvFilter over LevelFilter is
that you can do that...) or env vars to check, also the fallback behavior
is very constrained and opaque

so what we do instead is find the first of `{HUMMINGBIRD,RUST}_LOG` env
vars that's set. if its empty, we just use the default, which is now set
to `warn` for noisy deps and `info` for everything else. the reason why
we don't just check for the first non-empty is in case the user wants to
set `RUST_LOG="blah"` for some other rust tool but would like hummingbird
to run at default log level. this way, if they set `HUMMINGBIRD_LOG=""`
the other var gets ignored.

anyways this also includes some changes to how logs are formatted, here's
a sample:

```
   0.001781125s  INFO main hummingbird: Starting application
   0.219629708s  INFO playback hummingbird::playback::thread: Opened device: Ok("MacBook Pro Speakers"), format: Float32, rate: 44100, channel_count: 2
   0.219712875s  INFO playback hummingbird::playback::thread: Opening: "/Users/cam/Music/Tracks/10,000 Maniacs - Our Time in Eden/1-01 - 10,000 Maniacs - Noah’s Dove.ogg"
   0.283313583s  INFO  scanner hummingbird::library::scan: Scan complete, writing scan record and stopping
   0.283814250s  INFO  scanner hummingbird::library::scan: Scan record written to "/Users/cam/Library/Application Support/org.mailliw.hummingbird/scan_record.json"
   2.110743041s  INFO     main hummingbird::ui::global_actions: Quitting...
```

perhaps at some point we should add another tracing_subscriber layer that
writes to a file in data_dir at Level::DEBUG or something, possibly with
log rotation if it's simple enough to implement. want to check what other
rust apps do before i reinvent that wheel though